+++ src/main/scala/DataStructures/LightBlock.scala
@@ -4,7 +4,7 @@ import utils.Utils
 
 import scala.collection.mutable.ArrayBuffer
 
-case class LightBlock(id: Long, coords: (Int, Int), source: ArrayBuffer[SpatialEntity], targetIDs: ArrayBuffer[Long]) extends TBlock {
+case class LightBlock(id: Long, coords: (Int, Int), source: ArrayBuffer[SpatialEntity], targetIDs: ArrayBuffer[Int]) extends TBlock {
 
     /**
      * For each compaison in the block, return its id
@@ -17,7 +17,7 @@ case class LightBlock(id: Long, coords: (Int, Int), source: ArrayBuffer[SpatialE
         comparisonsIDs.toSet
     }
 
-    def getComparisonsPairs: ArrayBuffer[(Long, Long)]={
+    def getComparisonsPairs: ArrayBuffer[(Int, Int)]={
         val comparisonsPairs = for (s <-source; tid <- targetIDs)
             yield (s.id, tid)
         comparisonsPairs
@@ -27,14 +27,14 @@ case class LightBlock(id: Long, coords: (Int, Int), source: ArrayBuffer[SpatialE
      * For each compaison in the block, return its id
      * @return block's comparisons ids
      */
-    def getComparisons: ArrayBuffer[(Long, (SpatialEntity, Long))]={
+    def getComparisons: ArrayBuffer[(Long, (SpatialEntity, Int))]={
         for (s <-source; tID <- targetIDs)
             yield (Utils.bijectivePairing(s.id, tID), (s, tID))
     }
 
-    def getSourceIDs: ArrayBuffer[Long] =  source.map(se => se.id)
+    def getSourceIDs: ArrayBuffer[Int] =  source.map(se => se.id)
 
-    def getTargetIDs: ArrayBuffer[Long] = targetIDs.map(tid => tid)
+    def getTargetIDs: ArrayBuffer[Int] = targetIDs.map(tid => tid)
 
     def getSourceSize(): Long = source.size
 
@@ -43,7 +43,7 @@ case class LightBlock(id: Long, coords: (Int, Int), source: ArrayBuffer[SpatialE
 }
 
 object LightBlock{
-    def apply(coords: (Int, Int), source: ArrayBuffer[SpatialEntity], target: ArrayBuffer[Long]): LightBlock ={
+    def apply(coords: (Int, Int), source: ArrayBuffer[SpatialEntity], target: ArrayBuffer[Int]): LightBlock ={
         LightBlock(Utils.signedPairing(coords._1, coords._2), coords, source, target)
     }
 }
